mysql-Package Interfaces erstellt und Implementierung begonnen

+++ code/TwitterCrawler/src/test/DBcategorizerTest.java
@@ -73,8 +73,8 @@ public class DBcategorizerTest {
         try {
             DBtest t = new DBtest(access, log);
             t.sql("UPDATE accounts SET Categorized = 0 WHERE Id = 1;");
-            t.sql(" DELETE FROM accountCategory WHERE 1;");
-            t.sql("DELETE FROM category WHERE Name = \"testCP\";");
+            t.sql(" DELETE FROM accountCategory WHERE AccountId=1;");
+            // t.sql("DELETE FROM category WHERE Name = \"testCP\";");
         } catch (InstantiationException | IllegalAccessException
                 | ClassNotFoundException e) {
             e.printStackTrace();
@@ -99,8 +99,8 @@ public class DBcategorizerTest {
         try {
             DBtest t = new DBtest(access, log);
             t.sql("UPDATE accounts SET Categorized = 0 WHERE Id = 1;");
-            t.sql(" DELETE FROM accountCategory WHERE 1;");
-            t.sql("DELETE FROM category WHERE Name = \"testCP\";");
+            t.sql(" DELETE FROM accountCategory WHERE AccountId=1;");
+            // t.sql("DELETE FROM category WHERE Name = \"testCP\";");
         } catch (InstantiationException | IllegalAccessException
                 | ClassNotFoundException e) {
             e.printStackTrace();
@@ -122,8 +122,8 @@ public class DBcategorizerTest {
         try {
             DBtest t = new DBtest(access, log);
             t.sql("UPDATE accounts SET Categorized = 0 WHERE Id = 1;");
-            t.sql(" DELETE FROM accountCategory WHERE 1;");
-            t.sql("DELETE FROM category WHERE Name = \"testCP\";");
+            t.sql(" DELETE FROM accountCategory WHERE AccountId=1;");
+            // t.sql("DELETE FROM category WHERE Name = \"testCP\";");
             t.sql("DELETE FROM category WHERE Name = \"parent\";");
         } catch (InstantiationException | IllegalAccessException
                 | ClassNotFoundException e) {
@@ -149,8 +149,8 @@ public class DBcategorizerTest {
         try {
             DBtest t = new DBtest(access, log);
             t.sql("UPDATE accounts SET Categorized = 0 WHERE Id = 1;");
-            t.sql(" DELETE FROM accountCategory WHERE 1;");
-            t.sql("DELETE FROM category WHERE Name = \"testCP\";");
+            t.sql(" DELETE FROM accountCategory WHERE AccountId=1;");
+            // t.sql("DELETE FROM category WHERE Name = \"testCP\";");
             t.sql("DELETE FROM category WHERE Name = \"parent\";");
         } catch (InstantiationException | IllegalAccessException
                 | ClassNotFoundException e) {
